refactor: move all function to the utils folder
