+++ packages/openapi-parser/src/utils/escapeJsonPointer.test.ts
