+++ packages/openapi-parser/src/utils/getSegmentsFromPath.test.ts
