+++ packages/openapi-parser/src/utils/getSegmentsFromPath.ts
