+++ packages/openapi-parser/src/utils/index.ts
@@ -1,14 +1,20 @@
 export * from './details'
+export * from './escapeJsonPointer'
 export * from './filter'
 export * from './getEntrypoint'
 export * from './getListOfReferences'
+export * from './getSegmentsFromPath'
 export * from './isFilesystem'
+export * from './isObject'
 export * from './loadFiles'
 export * from './normalize'
 export * from './resolve'
+export * from './resolveReferences'
 export * from './toJson'
 export * from './toYaml'
+export * from './transformErrors'
 export * from './traverse'
+export * from './unescapeJsonPointer'
 export * from './upgrade'
 export * from './upgradeFromThreeToThreeOne'
 export * from './upgradeFromTwoToThree'
