+++ packages/openapi-parser/src/utils/unescapeJsonPointer.test.ts
